requirements.txt
sagemaker==2.232.1
boto3==1.35.25
duckdb==1.1.0
joblib==1.4.2
pylint==3.3.0
pyyaml==6.0.2
scikit-learn==1.5.2
xgboost==2.1.1